DTIMING  and Minor fixed in Painter and Oci8
git-svn-id: svn://ultimatepp.org/upp/trunk@878 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Oracle/Oci8.cpp
@@ -1062,8 +1062,8 @@ Vector<String> Oracle8::EnumReservedWords()
 
 void OracleBlob::SetStreamSize(int64 pos) {
 	ASSERT(pos <= GetStreamSize());
-	if(pos < GetStreamSize())
-		session->oci8.OCILobTrim(session->svchp, session->errhp, locp, pos);
+	if(pos < (int)GetStreamSize())
+		session->oci8.OCILobTrim(session->svchp, session->errhp, locp, (int)pos);
 }
 
 dword OracleBlob::Read(int64 at, void *ptr, dword size) {
